Preempt

+++ awkernel_lib/src/arch/aarch64/context/callee_saved_regs.rs
@@ -113,7 +113,7 @@ impl crate::context::Context for Context {
 
     unsafe fn set_entry_point(&mut self, entry: extern "C" fn(usize) -> !, arg: usize) {
         self.gp_regs.x19 = arg as u64;
-        self.gp_regs.x20 = entry as u64;
+        self.gp_regs.x20 = entry as usize as u64;
         self.gp_regs.x30 = entry_point as *const () as u64;
     }
